+++ jobs_manager/manager.py
@@ -32,9 +32,9 @@ class JobsManager:
     Job from the queue will be started later, when _start_monitoring detect it.
     If jobs queue is empty, monitoring is waiting for new jobs.
     """
-    def __init__(self, db_conf, mem_conf, disp_conf,
+    def __init__(self, db_conn, mem_conf, disp_conf,
                  resolver_conf, user_conf):
-        self.db_conf = db_conf
+        self.db_conn = db_conn
         self.mem_conf = mem_conf
         self.disp_conf = disp_conf
         self.r_conf = resolver_conf
@@ -55,7 +55,7 @@ class JobsManager:
         """
         try:
             job = Job(request=request,
-                      db_conf=self.db_conf,
+                      db_conn=self.db_conn,
                       mem_conf=self.mem_conf,
                       resolver_conf=self.r_conf,
                       tracker_max_interval=self.tracker_max_interval,
@@ -77,7 +77,7 @@ class JobsManager:
         :return:            results of checking job
         """
         job = Job(request=request,
-                  db_conf=self.db_conf,
+                  db_conn=self.db_conn,
                   mem_conf=self.mem_conf,
                   resolver_conf=self.r_conf,
                   tracker_max_interval=self.tracker_max_interval,